% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson_to_csv.R
\name{gojson2csv}
\alias{gojson2csv}
\title{geojson file to csv for strata description}
\usage{
gojson2csv(fn, vrep = "StratumNam", valrep = c("1511", "1512"), outfn)
}
\arguments{
\item{fn}{path to geojson files}

\item{vrep}{name of the column containing the Stratum names, default: "StratumNam"}

\item{valrep}{values that replace the values from val2rep}

\item{outfn}{path where the output csv should be saved}

\item{val2rep}{values to be replaced, all strata need to be in numeric format}
}
\description{
geojson file to csv for strata description
}
