% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson_to_csv.R, R/runapp.R
\name{planner}
\alias{planner}
\title{Add together two numbers}
\usage{
planner()

planner()
}
\arguments{
\item{fn}{path to geojson files}

\item{vrep}{name of the column containing the Stratum names, default: "StratumNam"}

\item{val2rep}{values to be replaced, all strata need to be in numeric format}

\item{valrep}{values that replace the values from val2rep}

\item{outfn}{path where the output csv should be saved}
}
\description{
Add together two numbers

Run the Survey Palnner app
}
